Separate server and client
Separate and simplify code structure: server & client.

Write `README.md` file to show how to run.
