+++ README.md
@@ -1,0 +1,19 @@
+# Chatting
+
+```bash
+npm install
+```
+
+## Run chatting server
+
+```bash
+npm run startServer
+```
+
+## Run chatting client
+
+```bash
+npm start
+```
+
+<http://localhost:8080/>
