+++ index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Chat app socket.io &amp; React</title>
+    <style> body { margin: 0; } </style>
+</head>
+<body>
+    <div id="root"></div>
+    <script type="text/javascript" src="./app.js"></script>
+</body>
+</html>
