+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "webpack-dev-server"
+    "start": "webpack-dev-server",
+    "startServer": "node server.js"
   },
   "author": "ZeroCho",
   "license": "MIT",
@@ -38,6 +39,7 @@
     "nodemon": "^1.19.0",
     "socket.io": "^2.3.0",
     "socket.io-client": "^2.3.0",
-    "webpack": "^4.30.0"
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.3.9"
   }
 }
