+++ src/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { createStore } from 'redux';
+
+import rootReducer from './reducers';
+import App from './App';
+
+const store = createStore(rootReducer);
+
+ReactDOM.render(
+    <Provider store={store}><App /></Provider>,
+    document.querySelector('#root')
+);
