+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import user from './user';
+
+const rootReducer = combineReducers({
+    user
+});
+
+export default rootReducer;
